1. 创建注销功能, 修复在页面无论点击那个user跳转info页面，显示的只有当前登录的user的info页面的bug 2. 优化project的add页面

+++ src/main/java/org/wang/web/service/impl/UserServiceImpl.java
@@ -78,7 +78,7 @@ public class UserServiceImpl implements UserService {
    * @return  DOCUMENT ME!
    */
   @Override public Integer save(User user) {
-    return userDao.save(user);
+    return (Integer)userDao.save(user);
   }
 
   //~ ------------------------------------------------------------------------------------------------------------------
